+++ src/common/POLFields/utils.js
@@ -10,3 +10,14 @@ export const isEresource = (format) => ERESOURCES.includes(format);
 export const isPhresource = (format) => PHRESOURCES.includes(format);
 
 export const isOtherResource = (format) => format === ORDER_FORMATS.other;
+
+export const isBinderyActiveDisabled = (orderFormat) => {
+  if (!orderFormat) {
+    return false;
+  }
+
+  return (
+    orderFormat !== ORDER_FORMATS.PEMix
+    && orderFormat !== ORDER_FORMATS.physicalResource
+  );
+};