+++ src/components/POLine/POLineDetails/POLineDetailsForm.test.js
@@ -137,4 +137,12 @@ describe('POLineDetailsForm', () => {
     expect(claimingIntervalField).toHaveValue(42);
     expect(claimingIntervalField).toBeValid();
   });
+
+  it('should `Bindary Active` field be disabled if "orderFormat" is not `physicalResource` nor `PEMix`', async () => {
+    renderPOLineDetailsForm(null, { orderFormat: 'Electronic' });
+
+    const binderyActiveField = screen.getByRole('checkbox', { name: 'ui-orders.poLine.isBinderyActive' });
+
+    expect(binderyActiveField).toBeDisabled();
+  });
 });