+++ src/settings/InventoryRecordTypeSelectField.js
@@ -1,11 +1,15 @@
-import React from 'react';
 import PropTypes from 'prop-types';
 import { FormattedMessage } from 'react-intl';
 
 import { FieldSelect, FieldSelectFinal } from '@folio/stripes-acq-components';
+
 import { INVENTORY_RECORDS_TYPE_FOR_SELECT } from '../components/POLine/const';
 
-const InventoryRecordTypeSelectField = ({ label, isRedux, ...rest }) => {
+const InventoryRecordTypeSelectField = ({
+  label,
+  isRedux,
+  ...rest
+}) => {
   const FieldComponent = isRedux ? FieldSelect : FieldSelectFinal;
 
   return (
@@ -23,6 +27,7 @@ InventoryRecordTypeSelectField.propTypes = {
   name: PropTypes.string.isRequired,
   disabled: PropTypes.bool,
   isRedux: PropTypes.bool,
+  onChange: PropTypes.func,
 };
 
 export default InventoryRecordTypeSelectField;