+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 * Create routing list modal. Refs UIOR-1191.
 * Add Users to routing list. Refs UIOR-1111.
 * Prioritize users in a routing list by drag and drop list. Refs UIOR-1112.
+* Create "Bindery active" flag in POL and order templates. Refs UIOR-1211.
 
 ## [6.0.4](https://github.com/folio-org/ui-orders/tree/v6.0.4) (2024-04-25)
 [Full Changelog](https://github.com/folio-org/ui-orders/compare/v6.0.3...v6.0.4)